+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+* generic client: ErrorFromResponse and NewHTTPError to allow easier mocking (#118, @LittleFox94)
+
 ### Changed
 * **breaking**: renamed corev1.Info to corev1.Resource (#113, @LittleFox94)