+++ CHANGELOG.md
@@ -11,10 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * clouddns/v1: creating a Record didn't retrieve its Identifier (#120, @LittleFox94)
 
 ### Added
-* generic client: FilterRequestURLHook for modifying request URLs
+* generic client: FilterRequestURLHook for modifying request URLs (#123, @marioreggiori)
 
 ### Changed
-* core/v1: ResourceWithTag uses RequestBodyHook and FilterRequestURLHook instead of RequestFilterHook
+* core/v1: ResourceWithTag uses RequestBodyHook and FilterRequestURLHook instead of RequestFilterHook (#123, @marioreggiori)
 
 <!--
 Please add your release notes under the correct category (Added, Changed, ...) and use the following format as a