+++ CHANGELOG.md
@@ -26,6 +26,10 @@ Some examples, more below in the actual changelog (newer entries are more likely
 ### Fixed
 * pkg/utils/object/compare.Reconcile now accepts arrays of `*struct` and `types.Object` as target/existing input (#145, @LittleFox94)
 
+### Changed
+* moved pkg/api.GetObjectIdentifier and related errors to pkg/api/types (#144, @LittleFox94)
+  - the previous locations are still available, but marked as deprecated
+
 ## [0.4.3] - 2022-05-04
 
 ### Fixed